Add files via upload
A simple notebooks that demonstrates how to read in an nd2 file and clean up some of the metadata to  be put in a dictionary.  May expand this into an nd2_reader that can be used to quickly read in an nd2 file to create sidpy dataset... however PIMS does a  pretty good job of this already.